# backend/internal/api/auth/login.go
package auth

import (
	"net/http"

	"github.com/gin-gonic/gin"

	"github.com/merynayr/mall-tenants/internal/logger"
	"github.com/merynayr/mall-tenants/internal/model"
	"github.com/merynayr/mall-tenants/internal/sys"
)

// Login аутентификация и получение JWT-токена
// @Summary Аутентификация
// @Description Аутентифицирует пользователя и возвращает JWT-токен
// @Tags user
// @Accept  json
// @Produce  json
// @Param body body model.AuthRequest true "Данные для аутентификации"
// @Success 200 {object} model.AuthResponse
// @Failure 400 {object} sys.ErrorResponse
// @Failure 401 {object} sys.ErrorResponse
// @Failure 500 {object} sys.ErrorResponse
// @Router /auth/login [post]
func (a *API) Login(c *gin.Context) {
	var req model.AuthRequest

	if err := c.ShouldBindJSON(&req); err != nil {
		logger.Debug(err.Error())
		sys.HandleError(c, sys.InvalidRequestError)
		return
	}

	authResponse, err := a.authService.Login(c.Request.Context(), req.Email, req.Password)
	if err != nil {
		sys.HandleError(c, err)
		return
	}

	a.setCookies(c, authResponse.RefreshToken, authResponse.AccessToken)

	c.JSON(http.StatusOK, authResponse)
}